Remove dependency on protobuf
Use JSON instead of protobuf for serializing the token.
Refactor the token-related code into multiple files for better readability.
