+++ README.markdown
@@ -8,6 +8,10 @@ The current implementation exposes two endpoints:
 - /authenticate: Handles token authentication requests coming from Kubernetes
 - /ldapAuth: Issues token to be used when interacting with the Kubernetes API
 
+Pre-requisites
+--------------
+- Certificate and corresponding private key for the webhook server
+- Certificate and corresponding private key for the Kubernetes webhook client
 
 Starting the webhook server
 ----------------
