--- token/proto/.gitignore
@@ -1,2 +1,0 @@
-*.pb.go 
-*.pb
