--- token/proto/doc.go
@@ -1,5 +1,0 @@
-// The pb package contains the definitions of the proto3 structures
-// used for representing tokens.
-//
-// TODO(dlg): actual documentation
-package pb
