--- token/proto/google/protobuf/.gitignore
@@ -1,1 +1,0 @@
-any.pb.go
